[WHD-221] Refactor: Add exception layer to RefreshToken repo
