+++ src/main/java/woohakdong/server/domain/refreshToken/RefreshToken.java
@@ -1,4 +1,4 @@
-package woohakdong.server.domain.refresh;
+package woohakdong.server.domain.refreshToken;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
