+++ src/main/java/woohakdong/server/domain/refreshToken/RefreshTokenJpaRepository.java
@@ -1,9 +1,9 @@
-package woohakdong.server.domain.refresh;
+package woohakdong.server.domain.refreshToken;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenJpaRepository extends JpaRepository<RefreshToken, Long> {
     Boolean existsByRefresh(String refresh);
 
     @Transactional
