+++ src/main/java/woohakdong/server/domain/refreshToken/RefreshTokenRepository.java
@@ -1,0 +1,9 @@
+package woohakdong.server.domain.refreshToken;
+
+public interface RefreshTokenRepository {
+    RefreshToken save(RefreshToken refreshToken);
+
+    Boolean existsByRefresh(String refresh);
+
+    void deleteByRefresh(String refresh);
+}
